Rename PCM-Java tests project
