+++ tests/tools.vitruv.applications.pcmjava.pojotransformations.editortests/src/tools/vitruv/applications/pcmjava/pojotransformations/editortests/java2pcm/InterfaceMappingTransformationTest.java
@@ -6,7 +6,7 @@ import org.palladiosimulator.pcm.repository.BasicComponent;
 import org.palladiosimulator.pcm.repository.OperationInterface;
 import org.palladiosimulator.pcm.repository.Repository;
 
-import tools.vitruv.applications.pcmjava.tests.pojotransformations.pcm2java.Pcm2JavaTestUtils;
+import tools.vitruv.applications.pcmjava.pcm2java.Pcm2JavaTestUtils;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
