--- tests/tools.vitruv.applications.pcmjava.pojotransformations.tests/src/tools/vitruv/applications/pcmjava/tests/pojotransformations/pcm2java/repository/CollectionDataTypeMappingTransformationTest.java
@@ -1,60 +1,0 @@
-package tools.vitruv.applications.pcmjava.tests.pojotransformations.pcm2java.repository;
-
-import org.junit.jupiter.api.Test;
-import org.palladiosimulator.pcm.repository.CollectionDataType;
-import org.palladiosimulator.pcm.repository.CompositeDataType;
-import org.palladiosimulator.pcm.repository.PrimitiveDataType;
-import org.palladiosimulator.pcm.repository.PrimitiveTypeEnum;
-import org.palladiosimulator.pcm.repository.Repository;
-
-import tools.vitruv.applications.pcmjava.tests.pojotransformations.pcm2java.Pcm2JavaTestUtils;
-import tools.vitruv.applications.pcmjava.tests.pojotransformations.pcm2java.Pcm2JavaTransformationTest;
-
-public class CollectionDataTypeMappingTransformationTest extends Pcm2JavaTransformationTest {
-
-	@Test
-	public void testAddCollectionDataTypeWithoutInnerType() throws Throwable {
-		final Repository repo = this.createAndSyncRepository(Pcm2JavaTestUtils.REPOSITORY_NAME);
-
-		this.getUserInteraction().addNextSingleSelection(0);
-		final CollectionDataType collectionDataType = this.addCollectionDatatypeAndSync(repo,
-				Pcm2JavaTestUtils.COLLECTION_DATA_TYPE_NAME, null);
-
-		this.assertDataTypeCorrespondence(collectionDataType);
-	}
-
-	@Test
-	public void testAddCollectionDataTypeWithPrimitiveTypeStringAsInnerType() throws Throwable {
-		final PrimitiveTypeEnum pte = PrimitiveTypeEnum.STRING;
-		this.testAddCollectionDataTypeWithPrimitiveInnerType(pte);
-	}
-
-	@Test
-	public void testAddCollectionDataTypeWithPrimitiveTypeIntAsInnerType() throws Throwable {
-		this.testAddCollectionDataTypeWithPrimitiveInnerType(PrimitiveTypeEnum.INT);
-	}
-
-	@Test
-	public void testAddCollectionDataTypeWithComplexInnerType() throws Throwable {
-		final Repository repo = this.createAndSyncRepository(Pcm2JavaTestUtils.REPOSITORY_NAME);
-		final CompositeDataType compositeDataType = this.createAndSyncCompositeDataType(repo);
-
-		this.getUserInteraction().addNextSingleSelection(0);
-		final CollectionDataType collectionDataType = this.addCollectionDatatypeAndSync(repo,
-				Pcm2JavaTestUtils.COLLECTION_DATA_TYPE_NAME, compositeDataType);
-
-		this.assertDataTypeCorrespondence(collectionDataType);
-	}
-
-	protected void testAddCollectionDataTypeWithPrimitiveInnerType(final PrimitiveTypeEnum pte) throws Throwable {
-		final Repository repo = this.createAndSyncRepository(Pcm2JavaTestUtils.REPOSITORY_NAME);
-		final PrimitiveDataType primitiveType = createPrimitiveDataType(pte, repo);
-
-		this.getUserInteraction().addNextSingleSelection(0);
-		final CollectionDataType collectionDataType = this.addCollectionDatatypeAndSync(repo,
-				Pcm2JavaTestUtils.COLLECTION_DATA_TYPE_NAME, primitiveType);
-
-		this.assertDataTypeCorrespondence(collectionDataType);
-	}
-
-}
