--- tests/tools.vitruv.applications.pcmjava.pojotransformations.tests/src/tools/vitruv/applications/pcmjava/tests/pojotransformations/pcm2java/repository/PcmParameterMappingTransformationTest.java
@@ -1,140 +1,0 @@
-package tools.vitruv.applications.pcmjava.tests.pojotransformations.pcm2java.repository;
-
-import org.eclipse.emf.ecore.util.EcoreUtil;
-import org.emftext.language.java.members.InterfaceMethod;
-import org.emftext.language.java.parameters.OrdinaryParameter;
-import org.junit.jupiter.api.Test;
-
-import org.palladiosimulator.pcm.repository.CompositeDataType;
-import org.palladiosimulator.pcm.repository.OperationSignature;
-import org.palladiosimulator.pcm.repository.Parameter;
-import org.palladiosimulator.pcm.repository.PrimitiveDataType;
-import org.palladiosimulator.pcm.repository.PrimitiveTypeEnum;
-import org.palladiosimulator.pcm.repository.Repository;
-
-import com.google.common.collect.Iterables;
-
-import tools.vitruv.applications.pcmjava.tests.pojotransformations.pcm2java.Pcm2JavaTestUtils;
-import tools.vitruv.applications.pcmjava.tests.pojotransformations.pcm2java.Pcm2JavaTransformationTest;
-import tools.vitruv.applications.util.temporary.pcm.PcmParameterUtil;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class PcmParameterMappingTransformationTest extends Pcm2JavaTransformationTest {
-
-	@Test
-	public void testAddParameter() throws Throwable {
-		final Parameter param = this.createAndSyncRepoOpSigAndParameter();
-
-		this.assertParameterCorrespondences(param);
-	}
-
-	@Test
-	public void testChangeParameterName() throws Throwable {
-		final Parameter param = this.createAndSyncRepoOpSigAndParameter();
-
-		PcmParameterUtil.setParameterName(param, Pcm2JavaTestUtils.PARAMETER_NAME + Pcm2JavaTestUtils.RENAME);
-		propagate();
-
-		this.assertParameterCorrespondences(param);
-	}
-
-	@Test
-	public void testChangeParameterType() throws Throwable {
-		final Parameter param = this.createAndSyncRepoOpSigAndParameter();
-
-		final Repository repo = param.getOperationSignature__Parameter().getInterface__OperationSignature()
-				.getRepository__Interface();
-		final PrimitiveDataType pdt = createPrimitiveDataType(PrimitiveTypeEnum.STRING, repo);
-		param.setDataType__Parameter(pdt);
-		propagate();
-
-		this.assertParameterCorrespondences(param);
-	}
-
-	@Test
-	public void testAddParameterWithCompositeDataType() throws Throwable {
-		final Parameter param = super.createAndSyncRepoOpSigAndParameterWithDataTypeName(
-				Pcm2JavaTestUtils.COMPOSITE_DATA_TYPE_NAME, Pcm2JavaTestUtils.PARAMETER_NAME);
-
-		this.assertParameterCorrespondences(param);
-	}
-
-	@Test
-	public void testRemoveParameter() throws Throwable {
-		final Parameter param = this.createAndSyncRepoOpSigAndParameter();
-
-		this.assertParameterCorrespondences(param);
-		final OperationSignature opSig = param.getOperationSignature__Parameter();
-		EcoreUtil.remove(param);
-		propagate();
-
-		this.assertCorrectSignatureMappingWithParameters(opSig, 0);
-	}
-
-	@Test
-	public void testAddMultipleParameters() throws Throwable {
-		final Parameter param = super.createAndSyncRepoOpSigAndParameterWithDataTypeName(
-				Pcm2JavaTestUtils.COMPOSITE_DATA_TYPE_NAME + "_1", "compositeParam1");
-		final OperationSignature opSig = param.getOperationSignature__Parameter();
-		final CompositeDataType cdt = super.createAndSyncCompositeDataType(
-				opSig.getInterface__OperationSignature().getRepository__Interface(),
-				Pcm2JavaTestUtils.COMPOSITE_DATA_TYPE_NAME + "_2");
-
-		final Parameter param2 = super.addAndSyncParameterWithPrimitiveTypeToSignature(opSig);
-		final Parameter param3 = super.addAndSyncParameterToSignature(opSig, cdt, "compositeParam2");
-
-		this.assertParameterCorrespondences(param);
-		this.assertParameterCorrespondences(param2);
-		this.assertParameterCorrespondences(param3);
-	}
-
-	@Test
-	public void testAddMultipleParametersAndRemoveOne() throws Throwable {
-		final Parameter param = super.createAndSyncRepoOpSigAndParameterWithDataTypeName(
-				Pcm2JavaTestUtils.COMPOSITE_DATA_TYPE_NAME + "_1", "compositeParam1");
-		final OperationSignature opSig = param.getOperationSignature__Parameter();
-		final CompositeDataType cdt = super.createAndSyncCompositeDataType(
-				opSig.getInterface__OperationSignature().getRepository__Interface(),
-				Pcm2JavaTestUtils.COMPOSITE_DATA_TYPE_NAME + "_2");
-
-		final Parameter param2 = super.addAndSyncParameterWithPrimitiveTypeToSignature(opSig);
-		final Parameter param3 = super.addAndSyncParameterToSignature(opSig, cdt, "compositeParam2");
-
-		this.assertCorrectSignatureMappingWithParameters(opSig, 3);
-		this.assertParameterCorrespondences(param);
-		this.assertParameterCorrespondences(param2);
-		this.assertParameterCorrespondences(param3);
-
-		EcoreUtil.remove(param2);
-		propagate();
-		this.assertCorrectSignatureMappingWithParameters(opSig, 2);
-	}
-
-	/**
-	 * assert that the parameter type has a correspondence to the type of the
-	 * parameter in the JaMoPP model
-	 */
-	@SuppressWarnings("unchecked")
-	private void assertParameterCorrespondences(final Parameter param) throws Throwable {
-		this.assertDataTypeCorrespondence(param.getDataType__Parameter());
-		this.assertCorrespondencesAndCompareNames(param, 1, new java.lang.Class[] { OrdinaryParameter.class },
-				new String[] { param.getParameterName() });
-	}
-
-	private void assertCorrectSignatureMappingWithParameters(final OperationSignature signature,
-			int expectedParameterCount) throws Throwable {
-		Iterable<InterfaceMethod> ims = getCorrespondingEObjects(signature, InterfaceMethod.class);
-		assertEquals(1, Iterables.size(ims));
-		InterfaceMethod im = ims.iterator().next();
-		assertEquals(expectedParameterCount, im.getParameters().size());
-		for (Parameter curParam : signature.getParameters__OperationSignature()) {
-			Iterable<OrdinaryParameter> javaParams = getCorrespondingEObjects(curParam, OrdinaryParameter.class);
-			assertEquals(1, Iterables.size(javaParams));
-			OrdinaryParameter javaParam = javaParams.iterator().next();
-			assertEquals(im.getParameters().get(signature.getParameters__OperationSignature().indexOf(curParam)),
-					javaParam);
-		}
-	}
-
-}
