+++ tests/tools.vitruv.applications.pcmjava.tests/tools.vitruv.applications.pcmjava.pojotransformations.tests/META-INF/MANIFEST.MF
@@ -1,0 +1,20 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Vitruv PCM-Java Application Tests
+Bundle-SymbolicName: tools.vitruv.applications.pcmjava.tests
+Automatic-Module-Name: tools.vitruv.applications.pcmjava.tests
+Bundle-Version: 0.3.0.qualifier
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Require-Bundle: org.apache.log4j,
+ com.google.guava,
+ org.junit.jupiter.api,
+ org.hamcrest.core,
+ org.emftext.language.java.resource.java,
+ edu.kit.ipd.sdq.commons.util.java,
+ tools.vitruv.testutils.vsum,
+ tools.vitruv.applications.util.temporary,
+ tools.vitruv.applications.pcmjava
+Export-Package: tools.vitruv.applications.pcmjava.pcm2java,
+ tools.vitruv.applications.pcmjava.pcm2java.repository,
+ tools.vitruv.applications.pcmjava.pcm2java.system
+Bundle-Vendor: vitruv.tools
