+++ tests/tools.vitruv.applications.pcmjava.tests/tools.vitruv.applications.pcmjava.pojotransformations.tests/src/tools/vitruv/applications/pcmjava/pcm2java/system/AssemblyContextMappingTransformationTest.java
@@ -1,0 +1,26 @@
+package tools.vitruv.applications.pcmjava.pcm2java.system;
+
+import org.junit.jupiter.api.Test;
+
+import org.palladiosimulator.pcm.core.composition.AssemblyContext;
+import org.palladiosimulator.pcm.repository.BasicComponent;
+import org.palladiosimulator.pcm.repository.Repository;
+import org.palladiosimulator.pcm.system.System;
+
+import tools.vitruv.applications.pcmjava.pcm2java.Pcm2JavaTestUtils;
+import tools.vitruv.applications.pcmjava.pcm2java.Pcm2JavaTransformationTest;
+
+public class AssemblyContextMappingTransformationTest extends Pcm2JavaTransformationTest {
+
+	@Test
+	public void testCreateAssemblyContext() throws Throwable {
+		final System system = super.createAndSyncSystem(Pcm2JavaTestUtils.SYSTEM_NAME);
+		final Repository repo = super.createAndSyncRepository(Pcm2JavaTestUtils.REPOSITORY_NAME);
+		final BasicComponent basicComponent = super.addBasicComponentAndSync(repo);
+
+		final AssemblyContext assemblyContext = this.createAndSyncAssemblyContext(system, basicComponent);
+
+		this.assertAssemblyContext(assemblyContext);
+	}
+
+}
